refactoring populate worksheet appeals and issues to util function
